rename integration.test.tsx to integ.test.tsx
