+++ apps/ledger-live-desktop/tsconfig.json
@@ -12,7 +12,8 @@
     "paths": {
       "~/*": ["src/*"],
       "tests/*": ["tests/*"],
-      "LLD/*": ["src/mvvm/*"]
+      "LLD/*": ["src/mvvm/*"],
+      "@ledgerhq/coin-cardano/*": ["../../libs/coin-modules/coin-cardano/src/*"]
     },
     "outDir": "dist",
     "typeRoots": ["index-types.d.ts", "./node_modules/@types", "./node_modules/@testing-library"]
