+++ apps/ledger-live-mobile/tsconfig.json
@@ -15,7 +15,8 @@
       "~/*": ["src/*"],
       "LLM/*": ["src/mvvm/*"],
       "@tests/*": ["__tests__/*"],
-      "@mocks/*": ["__mocks__/*"]
+      "@mocks/*": ["__mocks__/*"],
+      "@ledgerhq/coin-cardano/*": ["../../libs/coin-modules/coin-cardano/src/*"]
     }
   },
   "include": ["src/**/*", "e2e/**/*", "**/*.d.ts", "services/**/*"],
